+++ api/src/test/java/io/opentelemetry/trace/SpanBuilderTest.java
@@ -58,7 +58,7 @@ public class SpanBuilderTest {
           }
 
           @Override
-          public long getDroppedAttributeCount() {
+          public int getDroppedAttributeCount() {
             return 0;
           }
         });
