+++ sdk/src/test/java/io/opentelemetry/sdk/trace/RecordEventsReadableSpanTest.java
@@ -320,7 +320,7 @@ public class RecordEventsReadableSpanTest {
             }
 
             @Override
-            public long getDroppedAttributeCount() {
+            public int getDroppedAttributeCount() {
               return 0;
             }
           });
